PerfDataDecode - split out new header PerfEventInfo.h
The PerfDataDecode library's PerfDataFile.h header is for decoding a
perf.data file. It contains structs that define "stuff you need to know
about an event" (similar to ETW's `EVENT_RECORD`). These structs are
useful in cases other than decoding a perf.data file, so they shouldn't
be tied to PerfDataFile.h. Move them to their own header, where they can
stand on their own.

Proof that this is a good idea: note that EventFormatter.cpp no longer
needs to include PerfDataFile.h.
